wip - Always return an array of recipients for email.to
While this is a change that will break backward compatibility, it is
ultimately beneficial in that it will simplify the implementation of
`EmailProcessor`s, which can now simply loop over the recipients in
`email.to` and check to see if the address is relevant to their app.

This will also fix problems which may have been hidden when `email.to`
returns only one of potentially several addresses, in what may not have
been a valid format depending on an app's `Griddler.configuration.to`
setting.

* Convert all adapters to wrap `normalized_params[:to]` in an array,
  even for a single recipient.
* Add a feature spec for the integration between adapters and
  `Griddler::Email` which ensures that the parsed address is the same
  regardless of email service.