Deprecate :token option for config.to
I am deprecating the token option because with the possibility of
getting multiple recipients it becomes dangerous to operate only on the
token.

Take, for example, an email delivered to `topic.1@griddler.com` and to
`sam@looped-in-to-conversation.com`. With only tokens, the client
application would only see ['sam', 'topic.1'], which could cause
problems.

:token will remain available through the :hash option, which is now the
default option.