Force controller spec to use sendgrid adapter
spec/features/adapters_and_email_spec.rb:9 was intermittently leaking
the configuration across tests and causing the controller spec to think
it was using the mandrill or cloudmailin adapter. Instead it should
default to sendgrid, which matches up more closely to the sample email
params we're posting to the action.

Further changes:

* Move a method up in mandrill adapter spec for easier reading/comprehension
* Add :default adapter. :default => :sendgrid
* Move adapter reset to spec_helper