Remove unnecessary Griddler::Email#process calls
I'm not entirely sure why these were present, but the tests all pass
without the call. Probably a mistake made early on that was copied
whenever a new test was added.

Since it's not clear from this diff, #process gets the
Griddler.configuration.processor_class and sends it
Griddler.configuration.processor_method with self (the Griddler::Email
instance) as the argument.

This is the first step in moving the process logic out of
Griddler::Email and into the controller.