Remove all configuration of addresses
Previously users could configure the format in which they would receive
addresses in to, from, and cc fields of Griddler::Email objects. The
options were :token, :full, :email, :name, or :hash, which was a hash of
all other options.

Configuring :token was previously deprecated, and :hash has been the
default for :from and :cc for a while now. All of these values are
pretty easy to get back with Hash#[], so we're simply removing the
configuration logic and always returning the hash, which is the most
generally useful option.